3x optimisation (slimline) in processing streaming data through flumine   Before: 67,918,608 function calls (67913764 primitive calls) in 96.685 seconds   After:  22,045,096 function calls (22040254 primitive calls) in 28.429 seconds
This brings processing within lightweight mode speed :) For example when using just bflw:

Normal: 63,434,255 function calls (63430540 primitive calls) in 57.798 seconds
Lightweight: 7,524,609 function calls (7520893 primitive calls) in 8.824 seconds
Slimline (new): 17,560,815 function calls (17557100 primitive calls) in 14.956 seconds

This is possible by following a django style `lazy` object creation, the main bulk of processing is turning {priceSize} into <PriceSize> objects but this is pointless a lot of the time as the object is not used or can be handled as a dict instead.
The inclusion of slots further reduces the processing time as well as reducing memory.

This optimisation will improve normal streaming as well as backtesting, with more speed, less CPU + ram and minimal reduction in usability.

(For reference 1.170258201 was used, AUS racing)