Use smart-open for opening historical files
>smart_open is a drop-in replacement for Python's built-in open():
it can
>do anything open can (100% compatible, falls back to native open
>wherever possible), plus lots of nifty extra stuff on top.

The main bit of "nifty extra stuff" it is being used for here is for
transparent decompression of compressed files.

Previously in docs/performance.md smart-open was mention and a monkey
patch was provided:

    with patch("builtins.open", smart_open.open):
        framework.add_strategy(strategy)
        framework.run()

Anyone who was using this monkey patch will not have their workflow
distrupted by this change (as I have essentially hard-coded in what was
being patched in).

At the time of writing the latest release version of smart-open
is 6.3.0.

https://github.com/RaRe-Technologies/smart_open