Add [speed]
Previously requirements.txt/install_requires only required:

    betfairlightweight==2.17.3

which will not bring in the faster libraries that
betfairlightweight[speed] would bring in.

If you're running a standalone virtual environment for flumine then
you can easily miss out on these faster dependencies.

This adds a [speed] mode which simply depends on
betfairlightweight[speed] ensuring that the right versions of the
transitive dependencies are pulled in for the given version of bflw.